Time for unit testing!
Add a few test using the google testing framework. Also fix some bugs
related to duplicated surfaces on the geometry. To avoid insert the same
surface into the geometry, the operator== was defined on the Surface
class. The operator calls a pure virtual function that should be
implemented on Surface's childs.