Reaction class was coded and used with macro-xs calculations
A reaction is simply a functor that takes a particle as an argument and
change the particle's state. The reaction is sampled with the Sampler
class, which is a generic template that will be used to sample stuff
along the code.

Also, the keff algorithm was coded on the "main" and apparently is
working fine.